Add easy svg plugin
During a project i was not able to upload a font-file in the Divi editor. The font-file was ok, Divi was ok, found out it was a wordpress restriction. Offcourse this restriction can be changed through a function in PHP, and that's what easy svg does. It enabled me to upload the font-file and use it in the Divi theme